backend/requirements.txt
# ===== requirements.txt =====
fastapi==0.104.1
uvicorn[standard]==0.24.0
undetected-chromedriver==3.5.4
selenium==4.15.2
pydantic==2.5.0
python-dotenv==1.0.0
